requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0

# Embeddings
sentence-transformers>=2.2.0

# Similarity search
faiss-cpu>=1.7.3

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0

# Progress bars
tqdm>=4.65.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.20.0

# Testing
pytest>=7.2.0
pytest-cov>=4.0.0

# Code quality
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
